moved the app bar into own file
moved appbar to navbar
